Added stops and fixed bounds
Added STOPs to the arrays for morse code because otherwise the
loop would keep reading memory and would print out whatever the
compiler liked to put there. Also fixed bounds for the string and
char that make it so we do not write overtop of other dots, dashs,
and letters.